드디어 서버 additional data 구조완성 테스트 해봐야함.

+++ SelectModelServer/GameServer_Select/GameServer_Select/MemoryPooler.h
@@ -33,7 +33,7 @@ private:
 public:
 	struct MemoryBlock {
 		char* pData = nullptr;
-		int iDataSize = 0;
+		size_t iDataSize = 0;
 	};
 
 #pragma endregion
